requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
cx_Oracle==8.0.0
Django==3.1.1
django-celery-beat==2.0.0
django-timezone-field==4.0
dnspython3==1.15.0
flower==0.9.5
future==0.18.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
ipython==7.16.1
kombu==4.6.11
ldap3==2.6
monotonic==1.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
pymssql==2.1.4  # not compatible python 3.8
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
py-zabbix==1.1.7
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
